modeltrackr/tests/test_views.py
```python
import unittest
from modeltrackr import app, views
from modeltrackr.models import User, Model, Recording
from modeltrackr.services import AI_Service, Chaturbate_Service, Recording_Service, Notification_Service

class TestViews(unittest.TestCase):

    def setUp(self):
        self.app = app.test_client()
        self.app.testing = True

    def test_login_user(self):
        response = self.app.post('/login', data=dict(username='test', password='test'))
        self.assertEqual(response.status_code, 200)
        self.assertIn('user_logged_in', response.data)

    def test_logout_user(self):
        response = self.app.get('/logout')
        self.assertEqual(response.status_code, 200)
        self.assertIn('user_logged_out', response.data)

    def test_get_model_status(self):
        response = self.app.get('/model_status')
        self.assertEqual(response.status_code, 200)
        self.assertIn('model_updated', response.data)

    def test_start_recording(self):
        response = self.app.post('/start_recording', data=dict(model_id=1))
        self.assertEqual(response.status_code, 200)
        self.assertIn('recording_started', response.data)

    def test_end_recording(self):
        response = self.app.post('/end_recording', data=dict(recording_id=1))
        self.assertEqual(response.status_code, 200)
        self.assertIn('recording_ended', response.data)

if __name__ == "__main__":
    unittest.main()
```